update frontend add recover middleware
update frontend
add recover middleware

+++ models/models.go
@@ -17,15 +17,16 @@ type DayRoster struct {
 }
 
 // Product ...
+
 type Product struct {
-	ID       int    `json:"id"`
-	Name     string `json:"name"`
-	Qty      int    `json:"qty,string,omitempty"`
-	Price    int    `json:"price,string,omitempty"`
-	Cost     int    `json:"cost,string,omitempty"`
-	Code     int    `json:"code,string,omitempty"`
-	Catagory string `json:"catagory,omitempty"`
-	Imgurl   string `json:"Imgurl,omitempty"`
+	ID       int     `json:"id"`
+	Name     string  `json:"name"`
+	Qty      int     `json:"qty,string,omitempty"`
+	Price    float64 `json:"price,string,omitempty"`
+	Cost     int     `json:"cost,string,omitempty"`
+	Code     int     `json:"code,string,omitempty"`
+	Catagory string  `json:"catagory,omitempty"`
+	Imgurl   string  `json:"Imgurl,omitempty"`
 }
 
 // https://blog.auberginesolutions.com/how-i-handled-possible-null-values-from-database-rows-in-golang/
